Share _to_deepgram_url to parse booleans correctly
