+++ livekit-plugins/livekit-plugins-deepgram/livekit/plugins/deepgram/stt.py
@@ -43,7 +43,7 @@ from livekit.agents.types import (
 )
 from livekit.agents.utils import AudioBuffer, is_given
 
-from ._utils import PeriodicCollector
+from ._utils import PeriodicCollector, _to_deepgram_url
 from .log import logger
 from .models import DeepgramLanguages, DeepgramModels
 
@@ -749,26 +749,6 @@ def prerecorded_transcription_to_speech_event(
             for alt in dg_alts
         ],
     )
-
-
-def _to_deepgram_url(opts: dict, base_url: str, *, websocket: bool) -> str:
-    # don't modify the original opts
-    opts = opts.copy()
-    if opts.get("keywords"):
-        # convert keywords to a list of "keyword:intensifier"
-        opts["keywords"] = [
-            f"{keyword}:{intensifier}" for (keyword, intensifier) in opts["keywords"]
-        ]
-
-    # lowercase bools
-    opts = {k: str(v).lower() if isinstance(v, bool) else v for k, v in opts.items()}
-
-    if websocket and base_url.startswith("http"):
-        base_url = base_url.replace("http", "ws", 1)
-
-    elif not websocket and base_url.startswith("ws"):
-        base_url = base_url.replace("ws", "http", 1)
-    return f"{base_url}?{urlencode(opts, doseq=True)}"
 
 
 def _validate_model(
